fix debug for single files
